Add run script
The purpose of the run script is to call `module load slurm/blanca`,
which has to done before calling the launch script (which includes
SBATCH commands), and which can't be called through subprocess.call.
To do this, I added a new class attribute, `_run_script`, a new attr
to SbatchLauncher, `run_script_path`, and I overrode several methods.