Activate the root conda environment
This sets various environment variables; e.g., PROJ_DIR for basemap.

I've hardcoded the name of the root environment. It would be better if
WMT executors were installed in different environments rather than
different minicondas. When this happens, the name of the env should
be a variable, possibly set and read from wmt.cfg.